+++ docs/linode_api4/objects/models.rst
@@ -139,11 +139,3 @@ Volume Models
    :exclude-members: api_endpoint, properties, derived_url_path, id_attribute, parent_id_name
    :undoc-members:
    :inherited-members:
-
-Polling Classes
----------------
-
-.. automodule:: linode_api4.objects.polling
-   :members:
-   :undoc-members:
-   :inherited-members:
