+++ docs/linode_api4/polling.rst
@@ -1,0 +1,12 @@
+Event Polling
+==========
+
+This project exposes a framework for dynamically polling on long-running Linode Events.
+
+See the :doc:`Event Polling Guide<../guides/event_polling>` for more details.
+
+EventPoller class
+-------------------
+
+.. autoclass:: linode_api4.EventPoller
+   :members:
