+++ linode_api4/__init__.py
@@ -4,3 +4,4 @@ from linode_api4.errors import ApiError, UnexpectedResponseError
 from linode_api4.linode_client import LinodeClient
 from linode_api4.login_client import LinodeLoginClient, OAuthScopes
 from linode_api4.paginated_list import PaginatedList
+from linode_api4.polling import EventPoller
