+++ linode_api4/groups/polling.py
@@ -2,7 +2,7 @@ import polling
 
 from linode_api4.groups import Group
 from linode_api4.objects.account import Event
-from linode_api4.objects.polling import EventPoller, TimeoutContext
+from linode_api4.polling import EventPoller, TimeoutContext
 
 
 class PollingGroup(Group):
