+++ linode_api4/objects/__init__.py
@@ -17,4 +17,3 @@ from .tag import Tag
 from .object_storage import *
 from .lke import *
 from .database import *
-from .polling import *
